refactor: share the window-client readDirTree helper between test files
test/city-credit-http-mcp.test.ts and test/prepaid-credit-acceptance.test.ts
each carried an identical 11-line readDirTree, with a comment claiming
it read every .ts file when it actually read every file regardless of
extension. Extract one shared test/helpers/read-dir-tree.ts with a
comment that matches what it does.
